feat(PolicyDetails): added inline edits improved and refactored
